requirements.txt
ExifRead==3.3.1
pillow==11.2.1
